Use TOC for begginers guide, style TOC ul's
Use a single Table of Contents for the Beginner's Guide that in turn
includes the subfolders, with the 001-xxx document acting as the first
level of the toctree.
Display the Table of Contents using a `guide-toc` CSS class that makes
first-level entries be displayed as h2 elements.
Fix syntax warnings.